moved include files and  code refactored
