+++ sensing/autoware_imu_corrector/test/test_gyro_bias_scale_estimation_module.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gyro_bias_estimator.hpp"
+#include "autoware/imu_corrector/gyro_bias_estimator.hpp"
 
 #include <Eigen/Dense>
 #include <rclcpp/rclcpp.hpp>
